Refs #0. Remove other old files.
